Refactor jsonld parser to not inherit from lepl
Also add a basic test to make sure it works some
