+++ pymantic/parsers/__init__.py
@@ -2,9 +2,11 @@ __all__ = ['ntriples_parser', 'nquads_parser', 'turtle_parser', 'jsonld_parser']
 
 from .lepl import (
     nquads_parser,
-    jsonld_parser,
 )
 from .lark import (
     ntriples_parser,
     turtle_parser,
+)
+from .jsonld import (
+    jsonld_parser,
 )
