+++ pymantic/parsers/lepl/__init__.py
@@ -1,6 +1,5 @@
-__all__ = ['ntriples_parser', 'nquads_parser', 'turtle_parser', 'jsonld_parser']
+__all__ = ['ntriples_parser', 'nquads_parser', 'turtle_parser']
 
 from .ntriples import ntriples_parser
 from .nquads import nquads_parser
 from .turtle import turtle_parser
-from .jsonld import jsonld_parser
